+++ plugins/provisioners/chef/provisioner/chef_solo.rb
@@ -106,10 +106,11 @@ module VagrantPlugins
         def share_folders(root_config, prefix, folders)
           folders.each do |type, local_path, remote_path|
             if type == :host
-              root_config.vm.synced_folder(
-                local_path, remote_path,
-                :id =>  "v-#{prefix}-#{self.class.get_and_update_counter(:shared_folder)}",
-                :nfs => @config.nfs)
+              opts = {}
+              opts[:id] = "v-#{prefix}-#{self.class.get_and_update_counter(:shared_folder)}"
+              opts[:type] = @config.synced_folder_type if @config.synced_folder_type
+
+              root_config.vm.synced_folder(local_path, remote_path, opts)
             end
           end
         end