+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 DEPRECATIONS:
 
+  - provisioners/chef-solo: The "nfs" setting has been replaced by
+    `synced_folder_type`. The "nfs" setting will be removed in the next
+    version.
   - provisioners/puppet: The "nfs" setting has been replaced by
     `synced_folder_type`. The "nfs" setting will be removed in the next
     version.
@@ -12,6 +15,9 @@ FEATURES:
 
 IMPROVEMENTS:
 
+  - provisioners/chef-solo: New config `synced_folder_type` replaces the
+    `nfs` option. This can be used to set the synced folders the provisioner
+    needs to any type. [GH-2709]
   - provisioners/puppet: New config `synced_folder_type` replaces the
     `nfs` option. This can be used to set the synced folders the provisioner
     needs to any type. [GH-2709]